Implement adding book to api successfully
